Move class into separate file
